refactor windows mouse tracker
