+++ include/MouseTracker.hpp
@@ -1,42 +1,40 @@
 #pragma once
 
+#ifdef _WIN32
 #include <thread>
-#include <atomic>
-#include <memory>
-#include <cstdint>
-
-namespace sapf {
+#endif
 
 /**
- * MouseTracker interface for cross-platform mouse position tracking.
- * 
- * Provides a common interface for tracking mouse position across
- * different platforms (Windows, Linux, macOS).
+ * TODO: Convert to non-virtual
+ * Cross-platform mouse position tracking.
  */
 class MouseTracker {
 public:
-    virtual ~MouseTracker() = default;
+    MouseTracker();
+    ~MouseTracker();
 
     /**
      * Get the current mouse X position normalized between 0.0 and 1.0
      */
-    virtual float getMouseX() const = 0;
+    float getMouseX() const;
 
     /**
      * Get the current mouse Y position normalized between 0.0 and 1.0
      * (0 is bottom, 1 is top of screen)
      */
-    virtual float getMouseY() const = 0;
+    float getMouseY() const;
+private:
 
-    /**
-     * Get the current mouse button state (true if pressed)
-     */
-    virtual bool getMouseButton() const = 0;
-
-    /**
-     * Create a platform-specific MouseTracker instance
-     */
-    static std::unique_ptr<MouseTracker> create();
+#ifdef _WIN32
+    void trackMouse();
+    
+    const int mScreenWidth;
+    const int mScreenHeight;
+    const float mRscreenWidth;
+    const float mRscreenHeight; 
+    std::atomic<bool> mRunning{false};
+    std::thread mTrackingThread;
+    std::atomic<float> mMouseX{0.0f};
+    std::atomic<float> mMouseY{0.0f};
+#endif
 };
-
-} // namespace sapf
